GUS-3834 Prevent circular reference
